pyproject.toml
[tool.poetry]
name = "individual-task-buffed"
version = "0.1.0"
description = ""
authors = ["uandriyko <uandriyko@github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
motor = "^3.3.2"
environs = "^9.5.0"
bcrypt = "^4.0.1"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
